tttttt

+++ AceCook/Repositories/OrderRepository.cs
@@ -275,7 +275,7 @@ namespace AceCook.Repositories
                 {
                     Status = g.Key,
                     Count = g.Count(),
-                    TotalValue = g.Sum(o => o.CtDhs.Sum(ct => ct.SoLuong * ct.DonGia ?? 0))
+                    TotalValue = g.Sum(o => o.CtDhs.Sum(ct => (decimal)(ct.SoLuong * ct.DonGia ?? 0)))
                 })
                 .OrderByDescending(x => x.Count)
                 .ToList();
@@ -302,7 +302,7 @@ namespace AceCook.Repositories
                 {
                     CustomerName = g.Key.TenKh ?? "Không xác định",
                     OrderCount = g.Count(),
-                    TotalValue = g.Sum(o => o.CtDhs.Sum(ct => ct.SoLuong * ct.DonGia ?? 0))
+                    TotalValue = g.Sum(o => o.CtDhs.Sum(ct => (decimal)(ct.SoLuong * ct.DonGia ?? 0)))
                 })
                 .OrderByDescending(x => x.TotalValue)
                 .Take(top)
